Fix up the generated usage pages.
- improve some brief descriptions
- fix cases where the full help did not actually get generated correctly
- remove cherrycSMT.jar, since it's really a library, not a callable program
- rename Portage->PortageII
- <prog> -h shows the program name as <prog> for all programs, not ./<prog> or
  the full path to <prog>, as sometimes happened.
- removed mteval2ospl.pl, mteval2ospl.sed, and ospl2mteval.pl, obsolete and not
  relevant to PortageII.  Their help messages required fixing, and I'm willing
  to fix obsolete stuff.

PortageLive:
- default installation directory is now /opt/PortageII

+++ src/tp_models/ptable.encode-scores.cc
@@ -217,7 +217,9 @@ process_line(TMEntry& entry, ostream& prelimScoresFile,
       // the alignment information for the given phrase pair.
       prelimScoresFile.write(reinterpret_cast<char*>(&prelimAlignmentPosn),sizeof(prelimAlignmentPosn));
       if (debug_scr_file)
-         fprintf(stderr, "SCR size_t: %lu (\\x%016lx)\n", prelimAlignmentPosn, prelimAlignmentPosn);
+         fprintf(stderr, "SCR size_t: %lu (\\x%016lx)\n",
+            static_cast<long unsigned int>(prelimAlignmentPosn),
+            static_cast<long unsigned int>(prelimAlignmentPosn));
 
       /*
       if (!al || !*al) {
